Manage resources for the migration to HTCondor 23 in a controlled manner
Add a script (htcondor_migration.py) that reads the resource definition from resources.yaml and produces a new resource definition in which a configurable fraction of the cluster resources are allocated to the secondary HTCondor cluster.

Inject the functionality of the script into synchronize.py so that the fraction allocated to the secondary cluster can be controlled from Jenkins via an optional command line argument that defaults to zero.