Change hera2 training start date to 2025-10-27
The person who requested the hera2 training noticed that the starting date on the confirmation email we sent them (2025-10-28) with subject "UseGalaxy.eu TIaaS request Approved, Training ID: hera2" does not match the starting date they requested (2025-10-27).

Indeed, according to the TIaaS admin site, this training starts on 2025-10-27 and finishes on 2025-10-28, but the starting date in resources.yaml is 2025-10-27.

Set the correct start date in resources.yaml.