Decrease resources available to training scholtz on 2025-10-27
On 2025-10-27, there is no compute capacity for all trainings as the resources.yaml file stands now. Decrease the amount of compute nodes allocated to the scholtz training (it lasts just for that day, and it is the one with the least participants during that date) from 2 to 1, to resolve capacity conflicts.